requirement.txt
py-trello==0.18.0
aiogram==2.13
beautifulsoup4==4.9.3
pysqlite3==0.4.6
threaded==4.1.0
lxml==4.6.3
pyserial==3.5
subprocess.run==0.0.8
webbot==0.34
pyshorteners==1.0.1
python-vlc==3.0.12118
pydub==0.25.1
wikipedia==1.4.0
deimport==0.3.2
torch==1.10.0
torchvision==0.11.1
torchaudio==0.10.0